+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Yellowbox Changelog
+## 0.8.0
+### Added
+* added the `Aerospike` extra, to set up an aerospike enterprise container.
 ## 0.7.9
 ### Fixed
 * Fixed issue in webserver where if a client connection would disconnect abrubtly, the entire